+++ app/javascript/components/index.jsx
@@ -1,3 +1,4 @@
 export * from './Company/Company';
 export * from './Dropdown/Dropdown';
+export * from './ExcludeRules/ExcludeRules';
 export * from './Repository/Repository';